changed some DWORD to Color

+++ project/source/core/Color.h
@@ -22,7 +22,7 @@ struct Color
 	bool operator == (const Color& c) const { return value == c.value; }
 	bool operator != (const Color& c) const { return value != c.value; }
 
-	operator DWORD () const { return (DWORD)value; }
+	operator uint () const { return value; }
 	operator Vec4 () const { return Vec4(float(r) / 255.f, float(g) / 255.f, float(b) / 255.f, float(a) / 255.f); }
 	operator Vec3 () const { return Vec3(float(r) / 255.f, float(g) / 255.f, float(b) / 255.f); }
